# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Task Manager API for incidents, service requests, work notes, and resolutions
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemAction } from './activityItemAction';
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  action: ActivityItemAction;
  title: string;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  priority?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}