# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Daily Task Manager API for incidents, service requests, work notes, and resolutions
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  total: number;
  open: number;
  in_progress: number;
  resolved: number;
  closed: number;
  critical: number;
  high: number;
}